Creating a 'callable' way to perform tests and runs
Since crontab-e might be a solution, having a 'via-terminal' way to perform tests seems nice
