+++ services/initial_clean.py
@@ -1,55 +1,6 @@
 import numpy
 import pandas
-from variables import VARIABLES
-
-
-def _is_able_to_parse() -> bool:
-    """
-    Verifies if data_file_path and test_file_path are able to parse through
-    pandas.read_csv
-    @rtype: bool
-    @return: a boolean representing parse viability
-    """
-    data_file_path = VARIABLES["data_file_path"]
-    test_file_path = VARIABLES["test_file_path"]
-    try:
-        pandas.read_csv(data_file_path,
-                        sep=',',
-                        header=None,
-                        skiprows=VARIABLES["data_file_skip_row"], )
-        pandas.read_csv(test_file_path,
-                        sep=',',
-                        header=None,
-                        skiprows=VARIABLES["test_file_skip_row"], )
-        return True
-    except pandas.errors.ParserError:
-        return False
-
-
-def _number_of_columns_is(expected_number_of_columns: int) -> bool:
-    """
-    Verifies if data_file_path and test_file_path has the expected number
-    of columns
-    @param expected_number_of_columns: an integer representing expected number
-    of columns
-    @type expected_number_of_columns: int
-    @rtype: bool
-    @return: a boolean representing the data_file_path and test_file_path
-    expected number of columns conformity
-    """
-    data_file_path = VARIABLES["data_file_path"]
-    test_file_path = VARIABLES["test_file_path"]
-    data_file_dataframe = pandas.read_csv(data_file_path,
-                                          sep=',',
-                                          header=None,
-                                          skiprows=VARIABLES["data_file_skip_row"], )
-    test_file_dataframe = pandas.read_csv(test_file_path,
-                                          sep=',',
-                                          header=None,
-                                          skiprows=VARIABLES["test_file_skip_row"], )
-    data_number_of_columns = len(data_file_dataframe.columns)
-    test_number_of_columns = len(test_file_dataframe.columns)
-    return data_number_of_columns == expected_number_of_columns and test_number_of_columns == expected_number_of_columns
+from .variables import VARIABLES
 
 
 def _concatenate_files() -> pandas.DataFrame:
@@ -101,7 +52,7 @@ def _initial_clean_process(dataframe: pandas.DataFrame):
     dataframe.to_csv("data/output.csv", index=False)
 
 
-def _initial_clean():
+def initial_clean():
     """
     Concatenates data_file_path and test_file_path (_concatenate_files)
     and converts all fields on {dataframe} to string, them drops lines where
